:fire: :tractor: - Renaming HomeController to LogoutController and removing the homepage that is not being used.
