+++ src/Employer/Employer.Web/Configuration/Routing/RouteNames.cs
@@ -25,9 +25,8 @@
 
         public const string EmployerDetails_Index_Get = "EmployerDetails_Index_Get";
         public const string EmployerDetails_Index_Post = "EmployerDetails_Index_Post";
-
-        public const string Home_Index_Get = "Home_Index_Get";
-        public const string Home_Logout_Get = "Home_Logout_Get";
+        
+        public const string Logout_Get = "Logout_Get";
 
         public const string Employer_Get = "Employer_Get";
         public const string Employer_Post = "Employer_Post";
