+++ src/Employer/Employer.Web/Views/Shared/_Layout.cshtml
@@ -96,7 +96,7 @@
                                     <li><a href="@ExternalLinks.Notifications" class="sub-menu-item">Notifications settings</a></li>
                                 </ul>
                             </li>
-                            <li><a asp-route="Home_Logout_Get" role="menuitem" class="menu-main">Sign out</a></li>
+                            <li><a asp-route="@RouteNames.Logout_Get" role="menuitem" class="menu-main">Sign out</a></li>
                         }
                         else
                         {
